pyproject.toml
[tool.poetry]
name = "data-engineering-project"
version = "0.1.0"
description = "data project"
authors = ["Milan Šmíd <smidmi@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = ">=2.2.3,<3.0.0"
numpy = ">=2.2.4,<3.0.0"
sqlalchemy = "<2.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
pyspark = ">=3.5.5,<4.0.0"
pytest = ">=8.3.5,<9.0.0"
apache-airflow = ">=2.10.5,<3.0.0"
python-dotenv = ">=1.1.0,<2.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
